split code to inject into internal/abi into a separate file
This way we can edit it as regular Go code, with the help of an LSP,
and we can also test and benchmark it.
